Refactor: KouluaineSelect component
